Implement optional controller for alternate control of Wizard
This new optional controller allows for developers to control
the wizard without the need for context or if the widget trying
to control the wizard is above the wizard's context.

This was implemented to allow Mobile AppBar's BackButton to
call Wizard's Back.

There is a bit of a workaround with this implementation because of
how WizardScope is tied to each generated Page. The controller
must check for "hasListeners" to ensure only one listener is added.

Otherwise without this check, each WizardScope will attach its
own controller listener. This has the drastic side effect of
multiple listeners calling next or back or whatever action is
triggered.

I can understand why the Wizard Router is designed this way
but it would probably make more sense to have one WizardScope
rather than several, and allow wizard scope to accept functions
rather than try to handle everything internally within the
Wizard class.

However this is a larger change, outside the scope of this
optional controller pattern.

There is a new file called actions.dart which demonstrates the use
and need of the controller.